e2e: packaged-desktop target — the real bundle attaches to the supervised daemon
The supervised-daemon attach path (ensureSupervisedConnection →
attachToSupervisedDaemon) and the bundled compiled sidecar only run when
app.isPackaged is true. Dev electron skips that branch and always spawns its
own desktop-sidecar, so the attach behaviour can't be proven against a dev
launch — the prior dev-electron attach test could never satisfy its own
assertion. Test the production artifact instead.

- electron-builder.e2e.config.ts: unsigned bundle (no notarize/hardenedRuntime,
  `dir` target) so the real package builds without Apple credentials; the
  release config is untouched.
- desktop-packaged vitest project + globalsetup: builds web UI → compiled
  sidecar → electron-vite → electron-builder, then publishes the launch exe and
  bundled sidecar path.
- desktop-packaged/supervised-attach.test.ts: starts the bundle's own compiled
  executor-sidecar as the supervised daemon, launches the packaged app, and
  asserts it attached (manifest still names the daemon's pid/kind, console
  renders through the bearer-gated daemon) rather than spawning a sidecar.
  Skips honestly where no GUI display is reachable instead of hanging on launch.
- Remove the superseded dev-electron attach test; fix a stale "Basic-auth"
  comment (the header is bearer).
